refactor(pages): extract useCopyCurrentUrl composable
The image- and album-detail pages each hand-rolled the same SSR-safe
copy-current-URL + 2s-notification logic. Extract it into
composables/useCopyCurrentUrl and have both pages consume it.
